fix renamed file import
